* Add order context and reference item use jsonblob type for item column of order, this is so when item price changes in the future, it would make it easier to maintain price history of existing orders * Fix prices type to decimal

+++ lib/mypos/product/item.ex
@@ -6,7 +6,7 @@ defmodule Mypos.Product.Item do
     field :added_on, :date
     field :description, :string
     field :name, :string
-    field :price, :integer
+    field :price, :decimal
     belongs_to(:category, Mypos.Product.Category)
 
     timestamps()
